fix max_features in gradient boosting

+++ AutoSklearn/components/classification/gradient_boosting.py
@@ -57,7 +57,7 @@ class GradientBoostingClassifier(AutoSklearnClassificationAlgorithm):
 
     def fit(self, X, Y):
         num_features = X.shape[1]
-        max_features = float(self.max_features) * (np.log(num_features) + 1)
+        max_features = int(float(self.max_features) * (np.log(num_features) + 1))
         # Use at most half of the features
         max_features = max(1, min(int(X.shape[1] / 2), max_features))
         self.estimator = sklearn.ensemble.GradientBoostingClassifier(
